src/main/java/br/eti/rslemos/bitsmagic/arithmetic/DelayedZeroIntOutputStream.java
/*******************************************************************************
 * BEGIN COPYRIGHT NOTICE
 * 
 * The MIT License (MIT)
 * 
 * Copyright (c) 2016 Rodrigo Lemos
 * 
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 * 
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 * 
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 * 
 * END COPYRIGHT NOTICE
 *******************************************************************************/
package br.eti.rslemos.bitsmagic.arithmetic;

import java.io.IOException;

import br.eti.rslemos.bitsmagic.stream.IntOutputStream;

class DelayedZeroIntOutputStream implements IntOutputStream {

	private final IntOutputStream stream;
	private long zeros;

	public DelayedZeroIntOutputStream(IntOutputStream stream) {
		this.stream = stream;
	}

	@Override public void writeInt(int v) throws IOException {
		if (!delayZero(v)) {
			drainZeros();
			stream.writeInt(v);
		}
	}

	private void drainZeros() throws IOException {
		for(; zeros != 0; zeros--)
			stream.writeInt(0);
	}

	private boolean delayZero(int v) {
		if (v == 0) {
			zeros++;
		
			if (zeros == 0)
				throw new IllegalStateException("Too many zeros");
		}
		
		return v == 0;
	}
	
}